Theme: Move ThemeOption codes into 'theme' module
