+++ app/src/main/java/soup/movie/MovieApplication.kt
@@ -3,7 +3,7 @@ package soup.movie
 import com.jakewharton.threetenabp.AndroidThreeTen
 import dagger.android.AndroidInjector
 import dagger.android.support.DaggerApplication
-import soup.movie.data.ThemeOptionManager
+import soup.movie.theme.ThemeOptionManager
 import soup.movie.di.DaggerApplicationComponent
 import soup.movie.notification.NotificationSpecs
 import javax.inject.Inject
