+++ app/src/main/java/soup/movie/di/ApplicationModule.kt
@@ -5,8 +5,10 @@ import dagger.Module
 import dagger.Provides
 import soup.movie.MovieApplication
 import soup.movie.analytics.EventAnalytics
-import soup.movie.data.ThemeOptionManager
+import soup.movie.theme.ThemeOptionManager
 import soup.movie.settings.impl.ThemeOptionSetting
+import soup.movie.theme.ThemeOption
+import soup.movie.theme.ThemeOptionStore
 import soup.movie.util.ImageUriProvider
 import javax.inject.Singleton
 
@@ -34,5 +36,14 @@ class ApplicationModule {
     @Provides
     fun provideThemeOptionManager(
         themeOptionSetting: ThemeOptionSetting
-    ): ThemeOptionManager = ThemeOptionManager(themeOptionSetting)
+    ): ThemeOptionManager = ThemeOptionManager(object : ThemeOptionStore {
+
+        override fun save(option: ThemeOption) {
+            themeOptionSetting.set(option)
+        }
+
+        override fun restore(): ThemeOption {
+            return themeOptionSetting.get()
+        }
+    })
 }
