+++ app/src/main/java/soup/movie/settings/impl/ThemeOptionSetting.kt
@@ -1,8 +1,8 @@
 package soup.movie.settings.impl
 
 import android.content.SharedPreferences
-import soup.movie.data.ThemeOptionManager.Companion.defaultThemeOption
-import soup.movie.data.model.ThemeOption
+import soup.movie.theme.ThemeOptionManager.Companion.defaultThemeOption
+import soup.movie.theme.ThemeOption
 import soup.movie.settings.PrefSetting
 
 class ThemeOptionSetting(
