+++ theme/src/main/java/soup/movie/theme/ThemeOptionStore.kt
@@ -1,0 +1,8 @@
+package soup.movie.theme
+
+interface ThemeOptionStore {
+
+    fun save(option: ThemeOption)
+
+    fun restore(): ThemeOption
+}
